refactor: move db dependency into router
