+++ backend/app/router/__init__.py
@@ -1,0 +1,3 @@
+from .books import router as books_router
+
+__all__ = ["books_router"]
